% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppActiveBinding-package.R
\docType{package}
\name{RcppActiveBinding-package}
\alias{RcppActiveBinding}
\alias{RcppActiveBinding-package}
\title{An 'Rcpp' interface to active bindings}
\description{
Provides an easy way to fill an environment with active bindings
that call a C++ function.
}
\details{
TBD
}

