% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindrcpp-package.R
\docType{package}
\name{bindrcpp-package}
\alias{bindrcpp}
\alias{bindrcpp-package}
\title{bindrcpp: An 'Rcpp' interface to active bindings}
\description{
Provides an easy way to fill an environment with active bindings
that call a C++ function.
}
\details{
TBD
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/bindrcpp}
  \item \url{https://krlmlr.github.io/bindrcpp}
  \item Report bugs at \url{https://github.com/krlmlr/bindrcpp/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

}

