% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppActiveBinding-package.r
\docType{package}
\name{RcppActiveBinding-package}
\alias{RcppActiveBinding}
\alias{RcppActiveBinding-package}
\title{What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\details{
TBD
}

